assigment 3

+++ conFusionServer/routes/dishRouter.js
@@ -21,6 +21,7 @@ dishRouter.route('/')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser, (req, res, next) => {
+	authenticate.verifyAdmin(req,next);
     Dishes.create(req.body)
     .then((dish) => {
         console.log('Dish Created ', dish);
@@ -31,10 +32,12 @@ dishRouter.route('/')
     .catch((err) => next(err));
 })
 .put(authenticate.verifyUser, (req, res, next) => {
+	authenticate.verifyAdmin(req,next);
     res.statusCode = 403;
     res.end('PUT operation not supported on /dishes');
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
+	authenticate.verifyAdmin(req,next);
     Dishes.remove({})
     .then((resp) => {
         res.statusCode = 200;
@@ -56,10 +59,12 @@ dishRouter.route('/:dishId')
     .catch((err) => next(err));
 })
 .post(authenticate.verifyUser, (req, res, next) => {
+	authenticate.verifyAdmin(req,next);
     res.statusCode = 403;
     res.end('POST operation not supported on /dishes/'+ req.params.dishId);
 })
 .put(authenticate.verifyUser, (req, res, next) => {
+	authenticate.verifyAdmin(req,next);
     Dishes.findByIdAndUpdate(req.params.dishId, {
         $set: req.body
     }, { new: true })
@@ -71,6 +76,7 @@ dishRouter.route('/:dishId')
     .catch((err) => next(err));
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
+	authenticate.verifyAdmin(req,next);
     Dishes.findByIdAndRemove(req.params.dishId)
     .then((resp) => {
         res.statusCode = 200;
@@ -129,6 +135,7 @@ dishRouter.route('/:dishId/comments')
         + req.params.dishId + '/comments');
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
+	authenticate.verifyAdmin(req,next);
     Dishes.findById(req.params.dishId)
     .then((dish) => {
         if (dish != null) {
@@ -210,16 +217,26 @@ dishRouter.route('/:dishId/comments/:commentId')
     .catch((err) => next(err));
 })
 .delete(authenticate.verifyUser, (req, res, next) => {
+	console.log(req);
     Dishes.findById(req.params.dishId)
     .then((dish) => {
         if (dish != null && dish.comments.id(req.params.commentId) != null) {
-            dish.comments.id(req.params.commentId).remove();
-            dish.save()
-            .then((dish) => {
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(dish);                
-            }, (err) => next(err));
+			if(dish.comments.id(req.params.commentId)._doc.author.equals(req.user._id))
+			{
+				dish.comments.id(req.params.commentId).remove();
+				dish.save()
+				.then((dish) => {
+					res.statusCode = 200;
+					res.setHeader('Content-Type', 'application/json');
+					res.json(dish);                
+				}, (err) => next(err));
+			}
+			else
+			{
+				err = new Error('You are not authorized to perform this operation!');
+				err.status = 403;
+				return next(err);
+			}
         }
         else if (dish == null) {
             err = new Error('Dish ' + req.params.dishId + ' not found');
